삭제 로직 구현
